+++ com.unity.netcode.gameobjects/Runtime/Core/NetworkManager.cs
@@ -44,7 +44,7 @@ namespace Unity.Netcode
         {
             get
             {
-                return NetworkConfig.SessionMode == SessionModeTypes.DistributedAuthority;
+                return NetworkConfig.NetworkTopology == NetworkTopologyTypes.DistributedAuthority;
             }
         }