+++ com.unity.netcode.gameobjects/TestHelpers/Runtime/IntegrationTestWithApproximation.cs
@@ -87,7 +87,7 @@ namespace Unity.Netcode.TestHelpers.Runtime
             return new Vector3(Random.Range(min, max), Random.Range(min, max), Random.Range(min, max));
         }
 
-        public IntegrationTestWithApproximation(SessionModeTypes sessionModeType) : base(sessionModeType) { }
+        public IntegrationTestWithApproximation(NetworkTopologyTypes networkTopologyType) : base(networkTopologyType) { }
 
         public IntegrationTestWithApproximation(HostOrServer hostOrServer) : base(hostOrServer) { }