+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkSpawnManagerTests.cs
@@ -52,7 +52,7 @@ namespace Unity.Netcode.RuntimeTests
         {
             if (m_DistributedAuthority)
             {
-                VerboseDebug($"Ignoring test: Clients have access to other player objects in {m_SessionModeType} mode.");
+                VerboseDebug($"Ignoring test: Clients have access to other player objects in {m_NetworkTopologyType} mode.");
                 return;
             }
             // client can't access server player
@@ -77,7 +77,7 @@ namespace Unity.Netcode.RuntimeTests
 
             if (!m_DistributedAuthority)
             {
-                VerboseDebug($"Ignoring test: Clients do not have access to other player objects in {m_SessionModeType} mode.");
+                VerboseDebug($"Ignoring test: Clients do not have access to other player objects in {m_NetworkTopologyType} mode.");
                 return;
             }
 
@@ -90,7 +90,7 @@ namespace Unity.Netcode.RuntimeTests
         {
             if (m_DistributedAuthority)
             {
-                VerboseDebug($"Ignoring test: Clients have access to other player objects in {m_SessionModeType} mode.");
+                VerboseDebug($"Ignoring test: Clients have access to other player objects in {m_NetworkTopologyType} mode.");
                 return;
             }