+++ testproject/Assets/Tests/Runtime/NetworkSceneManager/SceneEventProgressTests.cs
@@ -10,8 +10,8 @@ using Random = UnityEngine.Random;
 
 namespace TestProject.RuntimeTests
 {
-    [TestFixture(SessionModeTypes.DistributedAuthority)]
-    [TestFixture(SessionModeTypes.ClientServer)]
+    [TestFixture(NetworkTopologyTypes.DistributedAuthority)]
+    [TestFixture(NetworkTopologyTypes.ClientServer)]
     public class SceneEventProgressTests : NetcodeIntegrationTest
     {
         private const string k_SceneUsedToGetAsyncOperation = "EmptyScene";
@@ -25,7 +25,7 @@ namespace TestProject.RuntimeTests
         private List<ulong> m_ClientThatShouldNotHaveCompleted = new List<ulong>();
         private List<ulong> m_ClientThatShouldHaveCompleted = new List<ulong>();
 
-        public SceneEventProgressTests(SessionModeTypes sessionModeType) : base(sessionModeType) { }
+        public SceneEventProgressTests(NetworkTopologyTypes networkTopologyType) : base(networkTopologyType) { }
 
         private bool SceneEventProgressComplete(SceneEventProgress sceneEventProgress)
         {