+++ testproject/Assets/Tests/Runtime/SenderIdTests.cs
@@ -11,8 +11,8 @@ using UnityEngine.TestTools;
 namespace TestProject.RuntimeTests
 {
 
-    [TestFixture(SessionModeTypes.DistributedAuthority)]
-    [TestFixture(SessionModeTypes.ClientServer)]
+    [TestFixture(NetworkTopologyTypes.DistributedAuthority)]
+    [TestFixture(NetworkTopologyTypes.ClientServer)]
     public class SenderIdTests : NetcodeIntegrationTest
     {
         protected override int NumberOfClients => 2;
@@ -20,7 +20,7 @@ namespace TestProject.RuntimeTests
         private NetworkManager FirstClient => m_ClientNetworkManagers[0];
         private NetworkManager SecondClient => m_ClientNetworkManagers[1];
 
-        public SenderIdTests(SessionModeTypes sessionModeType) : base(sessionModeType) { }
+        public SenderIdTests(NetworkTopologyTypes networkTopologyType) : base(networkTopologyType) { }
 
         [UnityTest]
         public IEnumerator WhenSendingMessageFromServerToClient_SenderIdIsCorrect()